refactor(utils): split Bootstrap.ts, extracting DOM helpers into BootstrapHelpers
Move `checkWebGPUSupport`, `displayError`, `getCanvas`, `ErrorContent`,
and the default ID constants out of `Bootstrap.ts` into a new dedicated
`BootstrapHelpers.ts` module.

`Bootstrap.ts` now imports these from `BootstrapHelpers`, keeping its
responsibility focused on the demo lifecycle (DOM-ready wait,
WebGPU validation, canvas lookup, engine init).

Also replaces `container.innerHTML = ‘’` with a `removeChild` loop for
safer DOM mutation.
